audio: Non-Tunnel in bypass mode, AV out of Sync. [1/1]
PD#SWPL-178768

Problem:
At the beginning, because the avsync cases are too many
we choose the Tunnel + (Auto/PCM/Passthrough Mode) to
tune the avsync.
Secondly, DV follow this mothod to test tunnel on Exo.
Thirdly, ASDK test the Tunnel and Non-Tunnel with different
E-EDIDs, and this Kit can not cover the Non-Tunnel&Passthrough.
Fourthly, DV test 1.Exo with Tunnel+Non-Tunnel + (Auto/
PCM/Passthrough), 2.Mediaplayer with (Auto/PCM/Passthrough).
This time, the Non-Tunnel with Passthrough case, that AV
is out of sync, it comes to us.

Solution:
Tune the AVSync with Non-Tunnel&Passthrough case. For the DV
case, tune one V-sync time(16ms) for DV process.

Verify:
S6(S905X5)

Change-Id: Id98f9a0bdb359dd5f1f3dd334fbff8e830939576
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>