audio: stream is closing but wrongly used in different thread [1/1]
PD#SWPL-182192

Problem:
When the stream A is closing by adev_close_output_stream(),
the new stream B is built, and call mixer_main_buffer_write()
to close the stream-A-in-closing, then hit that crash.

Solution:
When stream A is closing, first step is to call standby().
need to check whether the stream_status is STREAM_STANDBY.

Verify:
Local test passed.

Change-Id: I857edfcb01a17d50589264c2348a00071762ecbb
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>