audio: fixed dtv ad output not smooth issue [1/1]
PD#SWPL-195694

Problem:
After Framing the package, the Framing velocity is
different between main and ad. If when find the main
dolby frame, but it did not find the ad dolby frame
and will write the mute ad frame, which will cause
ad output not smooth.

Solution:
1.Main data frame package which will Increase read speed.
2.AD path need not split frame.
3.A complete packet of data consists of a packet of
valid APTS and a number of packets with APTS of 0.
4.Last_AD_Esdata use to save the complete ad data package

Verify:
U t5d

Change-Id: I294eacb86b68524ba8c3339938f7fbbceb7fd2ec
Signed-off-by: binbin dai <binbin.dai@amlogic.com>